+++ projeto/backend/controllers/FaturaController.php
@@ -29,6 +29,7 @@ class FaturaController extends Controller
     /**
      * @inheritDoc
      */
+    // Método que permite definir o que o utilizador tem permissão para fazer
     public function behaviors()
     {
         return array_merge(
@@ -44,12 +45,7 @@ class FaturaController extends Controller
                     'class' => AccessControl::className(),
                     'rules' => [
                         [
-                            'actions' => ['view'],
-                            'allow' => true,
-                            'roles' => ['admin', 'funcionario', 'cliente'],
-                        ],
-                        [
-                            'actions' => ['index', 'create', 'update', 'delete', 'imprimir'],
+                            'actions' => ['index', 'create', 'update', 'delete', 'imprimir', 'view'],
                             'allow' => true,
                             'roles' => ['admin', 'funcionario'],
                         ],
@@ -64,6 +60,7 @@ class FaturaController extends Controller
      *
      * @return string
      */
+    // Método que vai para o index das faturas
     public function actionIndex()
     {
         $searchModel = new FaturaSearch();
@@ -86,67 +83,63 @@ class FaturaController extends Controller
      * @return string
      * @throws NotFoundHttpException if the model cannot be found
      */
+    // Método que vai para a view de uma fatura
     public function actionView($id)
     {
-        $fatura = $this->findModel($id);
+        if (\Yii::$app->user->can('viewFatura')) {
+            $fatura = $this->findModel($id);
 
-        $estabelecimento = Estabelecimento::find()->where(['id' => $fatura->estabelecimento_id])->one();
+            $estabelecimento = Estabelecimento::find()->where(['id' => $fatura->estabelecimento_id])->one();
 
-        $cliente = User::find()->where(['id' => $fatura->cliente_id])->one();
-        $perfilCliente = $cliente->profiles;
+            $cliente = User::find()->where(['id' => $fatura->cliente_id])->one();
+            $perfilCliente = $cliente->profiles;
 
+            $linhasFatura = LinhaFatura::find()->where(['fatura_id' => $id])->all();
 
-        $linhasFatura = LinhaFatura::find()->where(['fatura_id' => $id])->all();
+            $carrinho = CarrinhoCompra::find()->where(['fatura_id' => $id])->one();
 
-        $carrinho = CarrinhoCompra::find()->where(['fatura_id' => $id])->one();
+            $servicosIds = ArrayHelper::getColumn($linhasFatura, 'servico_id');
+            $servicos = Servico::find()->where(['id' => $servicosIds])->all();
 
-        $servicosids = ArrayHelper::getColumn($linhasFatura, 'servico_id');
-        $servicos = Servico::find()
-            ->where(['id' => $servicosids])
-            ->all();
+            $receitasIds = ArrayHelper::getColumn($linhasFatura, 'receita_medica_id');
+            $receitas = ReceitaMedica::find()->where(['id' => $receitasIds])->all();
 
-        $receitasids = ArrayHelper::getColumn($linhasFatura, 'receita_medica_id');
-        $receitas = ReceitaMedica::find()
-            ->where(['id' => $receitasids])
-            ->all();
+            if ($carrinho !== null) {
+                $carrinhoId = $carrinho->id;
+                $linhasCarrinho = LinhaCarrinho::find()->where(['carrinho_compra_id' => $carrinhoId])->all();
 
-        if ($carrinho !== null) {
-            $carrinhoId = $carrinho->id;
-            $linhasCarrinho = LinhaCarrinho::find()->where(['carrinho_compra_id' => $carrinhoId])->all();
-
-            $produtos = [];
-            foreach ($linhasCarrinho as $linhaCarrinho) {
-                $produto = Produto::find()->where(['id' => $linhaCarrinho->produto_id])->one();
-                if ($produto) {
-                    $produtos[] = $produto;
+                $produtos = [];
+                foreach ($linhasCarrinho as $linhaCarrinho) {
+                    $produto = Produto::find()->where(['id' => $linhaCarrinho->produto_id])->one();
+                    if ($produto) {
+                        $produtos[] = $produto;
+                    }
                 }
+
+                return $this->render('view', [
+                    'fatura' => $fatura,
+                    'estabelecimento' => $estabelecimento,
+                    'cliente' => $cliente,
+                    'perfilCliente' => $perfilCliente,
+                    'linhasFatura' => $linhasFatura,
+                    'servicos' => $servicos,
+                    'receitas' => $receitas,
+                    'linhasCarrinho' => $linhasCarrinho,
+                    'produtos' => $produtos,
+                ]);
+            } else {
+                return $this->render('view', [
+                    'fatura' => $fatura,
+                    'estabelecimento' => $estabelecimento,
+                    'cliente' => $cliente,
+                    'perfilCliente' => $perfilCliente,
+                    'linhasFatura' => $linhasFatura,
+                    'servicos' => $servicos,
+                    'receitas' => $receitas,
+                ]);
             }
-
-            return $this->render('view', [
-                'fatura' => $fatura,
-                'estabelecimento' => $estabelecimento,
-                'cliente' => $cliente,
-                'perfilCliente' => $perfilCliente,
-                'linhasFatura' => $linhasFatura,
-                'servicos' => $servicos,
-                'receitas' => $receitas,
-                'linhasCarrinho' => $linhasCarrinho,
-                'produtos' => $produtos,
-            ]);
         }
-        else
-        {
-            return $this->render('view', [
-                'fatura' => $fatura,
-                'estabelecimento' => $estabelecimento,
-                'cliente' => $cliente,
-                'perfilCliente' => $perfilCliente,
-                'linhasFatura' => $linhasFatura,
-                'servicos' => $servicos,
-                'receitas' => $receitas,
-            ]);
-        }
-
+        throw new NotFoundHttpException('Não tem permissões para aceder a esta página');
     }
 
     /**
@@ -154,56 +147,55 @@ class FaturaController extends Controller
      * If creation is successful, the browser will be redirected to the 'view' page.
      * @return string|\yii\web\Response
      */
+    // Método que permite criar uma nova fatura
     public function actionCreate()
     {
-        $fatura = new Fatura();
-        $estabelecimentos = Estabelecimento::find()->all();
-        $estabelecimentosItems = ArrayHelper::map($estabelecimentos, 'id', 'nome');
+        if (\Yii::$app->user->can('createCategorias')) {
+
+            $fatura = new Fatura();
+            $estabelecimentos = Estabelecimento::find()->all();
+            $estabelecimentosItems = ArrayHelper::map($estabelecimentos, 'id', 'nome');
 
 
-        $authManager = Yii::$app->authManager;
-        $clienteRole = $authManager->getRole('cliente');
-        $clientes = User::find()
-            ->innerJoin('auth_assignment', 'auth_assignment.user_id = user.id')
-            ->andWhere(['auth_assignment.item_name' => $clienteRole->name])
-            ->all();
-        $clientesItems = ArrayHelper::map($clientes, 'id', 'username');
+            $authManager = Yii::$app->authManager;
+            $clienteRole = $authManager->getRole('cliente');
+            $clientes = User::find()
+                ->innerJoin('auth_assignment', 'auth_assignment.user_id = user.id')
+                ->andWhere(['auth_assignment.item_name' => $clienteRole->name])
+                ->all();
+            $clientesItems = ArrayHelper::map($clientes, 'id', 'username');
 
-        if ($this->request->isPost) {
-            $perfilEmissor = Profile::findOne(['user_id' => Yii::$app->user->id]);
-            if (!$perfilEmissor) {
-                Yii::$app->session->setFlash('error', 'O utilizador não tem o perfil criado.');
-                return $this->redirect('index');
+            if ($this->request->isPost) {
+                $perfilEmissor = Profile::findOne(['user_id' => Yii::$app->user->id]);
+                if (!$perfilEmissor) {
+                    Yii::$app->session->setFlash('error', 'O utilizador não tem o perfil criado.');
+                    return $this->redirect('index');
+                }
+
+                $fatura->dta_emissao = date('Y-m-d');
+                $fatura->emissor_id = Yii::$app->user->id;
+                $fatura->valortotal = 0;
+                $fatura->ivatotal = 0;
+
+                if ($fatura->load($this->request->post()) && $fatura->save()) {
+                    $estabelecimento = $fatura->estabelecimento_id;
+                    $cliente = $fatura->cliente_id;
+                    return $this->redirect(['linhafatura/create', 'id_fatura' => $fatura->id, 'estabelecimento_id' => $estabelecimento, 'cliente' => $cliente]);
+                }
+            } else {
+                $fatura->loadDefaultValues();
             }
 
-            $fatura->dta_emissao = date('Y-m-d');
-            $fatura->emissor_id = Yii::$app->user->id;
-            $fatura->valortotal = 0;
-            $fatura->ivatotal = 0;
-
-            if ($fatura->load($this->request->post()) && $fatura->save()) {
-                $estabelecimento = $fatura->estabelecimento_id;
-                $cliente = $fatura->cliente_id;
-                return $this->redirect(['linhafatura/create', 'id_fatura' => $fatura->id, 'estabelecimento_id' => $estabelecimento, 'cliente' => $cliente]);
-            }
-        } else {
-            $fatura->loadDefaultValues();
+            return $this->render('create', [
+                'fatura' => $fatura,
+                'estabelecimento' => $estabelecimentosItems,
+                'cliente' => $clientesItems,
+            ]);
         }
-
-        return $this->render('create', [
-            'fatura' => $fatura,
-            'estabelecimento' => $estabelecimentosItems,
-            'cliente' => $clientesItems,
-        ]);
+        throw new NotFoundHttpException('Não tem permissões para aceder a esta página');
     }
 
-    /**
-     * Updates an existing Fatura model.
-     * If update is successful, the browser will be redirected to the 'view' page.
-     * @param int $id ID
-     * @return string|\yii\web\Response
-     * @throws NotFoundHttpException if the model cannot be found
-     */
+    // Método que permite editar uma fatura
     public function actionUpdate($id)
     {
         $fatura = Fatura::find()->where(['id' => $id])->one();
@@ -229,34 +221,23 @@ class FaturaController extends Controller
      * @return \yii\web\Response
      * @throws NotFoundHttpException if the model cannot be found
      */
+    // Método que permite apagar uma fatura
     public function actionDelete($id)
     {
-        $linhasFatura = LinhaFatura::find()->where(['fatura_id' => $id])->all();
-        foreach ($linhasFatura as $linhaFatura) {
-            $linhaFatura->delete();
-        }
-        $this->findModel($id)->delete();
+        if (\Yii::$app->user->can('deleteFatura')) {
+            $linhasFatura = LinhaFatura::find()->where(['fatura_id' => $id])->all();
+            foreach ($linhasFatura as $linhaFatura) {
+                $linhaFatura->delete();
+            }
+            $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+        }
+        throw new NotFoundHttpException('Não tem permissões para aceder a esta página');
     }
 
-    /**
-     * Finds the Fatura model based on its primary key value.
-     * If the model is not found, a 404 HTTP exception will be thrown.
-     * @param int $id ID
-     * @return Fatura the loaded model
-     * @throws NotFoundHttpException if the model cannot be found
-     */
-    protected function findModel($id)
-    {
-        if (($model = Fatura::findOne(['id' => $id])) !== null) {
-            return $model;
-        }
-
-        throw new NotFoundHttpException('The requested page does not exist.');
-    }
-
-    public function actionImprimir($id) // Ação para gerar o PDF da fatura
+    //Método que permite imprimir uma fatura
+    public function actionImprimir($id)
     {
         $fatura = $this->findModel($id);
 
@@ -269,11 +250,11 @@ class FaturaController extends Controller
 
         $carrinho = CarrinhoCompra::find()->where(['fatura_id' => $id])->one();
 
-        $servicosids = ArrayHelper::getColumn($linhasFatura, 'servico_id');
-        $servicos = Servico::find()->where(['id' => $servicosids])->all();
+        $servicosIds = ArrayHelper::getColumn($linhasFatura, 'servico_id');
+        $servicos = Servico::find()->where(['id' => $servicosIds])->all();
 
-        $receitasids = ArrayHelper::getColumn($linhasFatura, 'receita_medica_id');
-        $receitas = ReceitaMedica::find()->where(['id' => $receitasids])->all();
+        $receitasIds = ArrayHelper::getColumn($linhasFatura, 'receita_medica_id');
+        $receitas = ReceitaMedica::find()->where(['id' => $receitasIds])->all();
 
         if ($carrinho !== null) {
             $carrinhoId = $carrinho->id;
@@ -298,9 +279,7 @@ class FaturaController extends Controller
                 'linhasCarrinho' => $linhasCarrinho,
                 'produtos' => $produtos,
             ]);
-        }
-        else
-        {
+        } else {
             $content = $this->renderPartial('impressao', [
                 'fatura' => $fatura,
                 'cliente' => $cliente,
@@ -316,6 +295,22 @@ class FaturaController extends Controller
         $mpdf = new Mpdf();
         $mpdf->WriteHTML($content);
         $mpdf->Output('Fatura ' . $id . '.pdf', 'D');
+    }
 
+    /**
+     * Finds the Fatura model based on its primary key value.
+     * If the model is not found, a 404 HTTP exception will be thrown.
+     * @param int $id ID
+     * @return Fatura the loaded model
+     * @throws NotFoundHttpException if the model cannot be found
+     */
+    // Método que permite encontrar a fatura selecionada
+    protected function findModel($id)
+    {
+        if (($model = Fatura::findOne(['id' => $id])) !== null) {
+            return $model;
+        }
+
+        throw new NotFoundHttpException('The requested page does not exist.');
     }
 }
